Extract withOnyx types to a new file
